Add CLI entry-point functions and shortcuts
