+++ pyproject.toml
@@ -75,7 +75,9 @@ docs = [
 ]
 
 [project.scripts]
-biogtr = "sleap_nn.cli:cli"
+sleap-nn-train = "sleap_nn.train:main"
+sleap-nn-track = "sleap_nn.predict:main"
+sleap-nn-eval = "sleap_nn.evaluation:main"
 
 [project.urls]
 Homepage = "https://sleap.ai"
