+++ tests/tracking/candidates/test_fixed_window.py
@@ -1,6 +1,6 @@
 from typing import DefaultDict, Deque, List
 import numpy as np
-from sleap_nn.inference.predictors import run_inference
+from sleap_nn.predict import run_inference
 from sleap_nn.tracking.track_instance import TrackedInstanceFeature
 from sleap_nn.tracking.candidates.fixed_window import FixedWindowCandidates
 from sleap_nn.tracking.tracker import Tracker
